Adding support to OpenAI visual model
Now it's possible to paste the images from a clipboard with the context
menu, and ask GPT to analyze it.

It comes with a caveat. The OpenAI vision model does not support
function calling. Because of this the plugin needs to switch models
depending whether GPT vision capabilities are in use.